+++ messages/ko/base.php
@@ -85,7 +85,7 @@ return [
   'Select this setting for visible text nodes only. Uncheck this setting in case this element is used for example as HTML attribute value.' => '',
   'Settings' => '설정',
   'Show additional file information (size)' => '',
-  'Snipped-Layouts' => '',
+  'Snippet Layouts' => '',
   'Snippet' => '',
   'Snippets' => '',
   'Sort Order' => '정렬 순서',